Agentic Framework/pyproject.toml
[tool.poetry]
name = "ai-agent-framework"
version = "0.1.0"
description = "Production-ready AI agent framework for hackathons"
authors = ["AI Agent Framework Team <team@agents.dev>"]
readme = "README.md"
packages = [{include = "agents"}, {include = "core"}, {include = "communication"}, {include = "memory"}, {include = "config"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
google-generativeai = "^0.3.2"
aiohttp = "^3.9.1"
asyncio-mqtt = "^0.16.0"
redis = "^5.0.1"
loguru = "^0.7.2"
python-dotenv = "^1.0.0"
tenacity = "^8.2.3"
rich = "^13.7.0"
typer = "^0.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
pre-commit = "^3.6.0"
coverage = "^7.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["agents", "core", "communication", "memory", "config"]
omit = ["*/tests/*", "*/test_*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
] 